+++ internal/repositories/task/CHANGELOG.md
@@ -7,4 +7,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Make create unit test
+- Make create func `FindTaskByUser` and `AddTask`
 - Create file `task.go`